Tweak: Move tagName option to Advanced
